REF/TST: collect fillna tests
